Prototype/Prototype 2/arc-ai mini/arc-ai mini/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
pydantic==2.5.0
pyyaml==6.0.1
requests==2.31.0

